+++ example/urls.py
@@ -1,6 +1,11 @@
 from django.conf import settings
 from django.conf.urls import include, url
+from django.urls import path
+from django.views.generic import TemplateView
 from rest_framework import routers
+from rest_framework.schemas import get_schema_view
+
+from rest_framework_json_api.schemas.openapi import SchemaGenerator
 
 from example.views import (
     AuthorRelationshipView,
@@ -63,11 +68,21 @@ urlpatterns = [
     url(r'^authors/(?P<pk>[^/.]+)/relationships/(?P<related_field>\w+)$',
         AuthorRelationshipView.as_view(),
         name='author-relationships'),
+    path('openapi', get_schema_view(
+        title="Example API",
+        description="API for all things …",
+        version="1.0.0",
+        generator_class=SchemaGenerator
+    ), name='openapi-schema'),
+    path('swagger-ui/', TemplateView.as_view(
+        template_name='swagger-ui.html',
+        extra_context={'schema_url': 'openapi-schema'}
+    ), name='swagger-ui'),
 ]
-
 
 if settings.DEBUG:
     import debug_toolbar
+
     urlpatterns = [
         url(r'^__debug__/', include(debug_toolbar.urls)),
     ] + urlpatterns